# docs/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Blockchain Indexer API",
    "description": "A blockchain indexer service that processes blocks, manages balances, and provides rollback functionality",
    "version": "1.0.0",
    "contact": {
      "name": "API Support",
      "email": "support@blockchain-indexer.com"
    },
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
...
  "paths": {
...
    "/balance/{address}": {
      "get": {
        "summary": "Get address balance",
        "tags": [
          "Balance"
        ],
        "description": "Retrieve the current balance for a specific blockchain address",
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "path",
            "name": "address",
            "required": true,
            "description": "Blockchain address to query"
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "address": {
                      "type": "string"
                    },
                    "balance": {
                      "type": "number"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "error": {
                      "type": "string"
                    },
                    "address": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "error": {
                      "type": "string"
                    },
                    "address": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "503": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Development server"
    }
  ],
...
}